27/10/2022 party order api (client), create order note, post to vnpay

+++ route/admin.management.food.js
@@ -4,7 +4,7 @@ const verifyAdmin = require('../authentication/auth')
 const Food = require('../database/Food')
 const FoodType = require('../database/FoodType')
 const { uploadImage } = require('../function/driveAPI')
-const { checkText } = require('../function/Inspect')
+const { checkText, checkFoodImage } = require('../function/Inspect')
 
 router.get('/get-foods/:pageCur/:numOnPage', async (req, res) => {
     await new Food()
@@ -88,24 +88,44 @@ router.put('/food-edit', verifyAdmin, async (req, res) => {
             isExistFood = result
         })
     if (isExistFood) {
-        if (checkText(foodname)) {
-            if (checkText(fooddescription)) {
-                try {
-                    await new Food().updateFood(foodid, foodname, foodtype, fooddescription)
-                } catch (err) {
-                    return res.status(400).json({
-                        success: false,
-                        message: 'Tên món ăn đã trùng'
-                    });
-                }
-                //Cập nhật chi tiết khẩu phần món
-                if (Array.isArray(foodpriceration)) {
-                    await new Food().deleteAllFoodDetail(foodid)
+        if (!checkFoodImage(req.files.foodimage)) {
+            return res.status(400).json({
+                success: false,
+                message: 'Ảnh món ăn không phù hợp, chỉ hỗ trợ định dạng png, jpeg, jpg'
+            });
+        }
+        else
+            if (checkText(foodname)) {
+                if (checkText(fooddescription)) {
+                    try {
+                        await new Food().updateFood(foodid, foodname, foodtype, fooddescription)
+                    } catch (err) {
+                        return res.status(400).json({
+                            success: false,
+                            message: 'Tên món ăn đã trùng'
+                        });
+                    }
+                    //Cập nhật chi tiết khẩu phần món
+                    if (Array.isArray(foodpriceration)) {
+                        await new Food().deleteAllFoodDetail(foodid)
 
-                    foodpriceration.forEach(async e => {
-                        let foodpriceration = JSON.parse(e)
+                        foodpriceration.forEach(async e => {
+                            let foodpriceration = JSON.parse(e)
+                            try {
+                                await new Food().updateFoodDetail(foodid, foodpriceration.price, foodpriceration.ration)
+                            }
+                            catch (err) {
+                                return res.status(400).json({
+                                    success: false,
+                                    message: 'Có lỗi xảy ra khi cập nhật chi tiết khẩu phần món ăn'
+                                });
+                            }
+                        })
+                    } else if (foodpriceration) {
+                        await new Food().deleteAllFoodDetail(foodid)
+                        let foodpriceratione = JSON.parse(foodpriceration)
                         try {
-                            await new Food().updateFoodDetail(foodid, foodpriceration.price, foodpriceration.ration)
+                            await new Food().updateFoodDetail(foodid, foodpriceratione.price, foodpriceratione.ration)
                         }
                         catch (err) {
                             return res.status(400).json({
@@ -113,79 +133,66 @@ router.put('/food-edit', verifyAdmin, async (req, res) => {
                                 message: 'Có lỗi xảy ra khi cập nhật chi tiết khẩu phần món ăn'
                             });
                         }
-                    })
-                } else if (foodpriceration) {
-                    await new Food().deleteAllFoodDetail(foodid)
-                    let foodpriceratione = JSON.parse(foodpriceration)
+                    }
+                    //Cập nhật hình ảnh món
                     try {
-                        await new Food().updateFoodDetail(foodid, foodpriceratione.price, foodpriceratione.ration)
-                    }
-                    catch (err) {
-                        return res.status(400).json({
-                            success: false,
-                            message: 'Có lỗi xảy ra khi cập nhật chi tiết khẩu phần món ăn'
-                        });
-                    }
-                }
-                //Cập nhật hình ảnh món
-                try {
-                    if (Array.isArray(foodimagedeleted)) {
-                        foodimagedeleted.forEach(async e => {
-                            await new Food().deleteFoodImage(foodid, e)
-                        })
-                    } else if (foodimagedeleted) {
-                        await new Food().deleteFoodImage(foodid, foodimagedeleted)
-                    }
+                        if (Array.isArray(foodimagedeleted)) {
+                            foodimagedeleted.forEach(async e => {
+                                await new Food().deleteFoodImage(foodid, e)
+                            })
+                        } else if (foodimagedeleted) {
+                            await new Food().deleteFoodImage(foodid, foodimagedeleted)
+                        }
 
-                    if (Array.isArray(req.files.foodimage)) {
-                        req.files.foodimage.forEach(async (image, index) => {
+                        if (Array.isArray(req.files.foodimage)) {
+                            req.files.foodimage.forEach(async (image, index) => {
+                                let id
+                                await uploadImage(image)
+                                    .then(async (imageID) => {
+                                        id = imageID
+                                        await new Food().updateFoodImage(foodid, id, foodimagedescription[index])
+                                    })
+                            })
+                            return res.status(200).json({
+                                success: true,
+                                message: 'Cập nhật món ăn thành công, vui lòng đợi giây lát khi ảnh đang được tải lên'
+                            });
+                        } else if (req.files.foodimage) {
                             let id
-                            await uploadImage(image)
+                            await uploadImage(req.files.foodimage)
                                 .then(async (imageID) => {
                                     id = imageID
-                                    await new Food().updateFoodImage(foodid, id, foodimagedescription[index])
+                                    await new Food().updateFoodImage(foodid, id, foodimagedescription)
                                 })
-                        })
-                        return res.status(200).json({
-                            success: true,
-                            message: 'Cập nhật món ăn thành công, vui lòng đợi giây lát khi ảnh đang được tải lên'
-                        });
-                    } else if (req.files.foodimage) {
-                        let id
-                        await uploadImage(req.files.foodimage)
-                            .then(async (imageID) => {
-                                id = imageID
-                                await new Food().updateFoodImage(foodid, id, foodimagedescription)
-                            })
-                        return res.status(200).json({
-                            success: true,
-                            message: 'Cập nhật món ăn thành công, vui lòng đợi giây lát khi ảnh đang được tải lên'
-                        });
-                    } else {
-                        return res.status(200).json({
-                            success: true,
-                            message: 'Cập nhật món ăn thành công'
+                            return res.status(200).json({
+                                success: true,
+                                message: 'Cập nhật món ăn thành công, vui lòng đợi giây lát khi ảnh đang được tải lên'
+                            });
+                        } else {
+                            return res.status(200).json({
+                                success: true,
+                                message: 'Cập nhật món ăn thành công'
+                            });
+                        }
+                    } catch (err) {
+                        console.log(err)
+                        return res.status(400).json({
+                            success: false,
+                            message: 'Có lỗi xảy ra khi cập nhật hình ảnh món ăn'
                         });
                     }
-                } catch (err) {
-                    console.log(err)
+                } else {
                     return res.status(400).json({
                         success: false,
-                        message: 'Có lỗi xảy ra khi cập nhật hình ảnh món ăn'
+                        message: 'Mô tả của món ăn không hợp lệ'
                     });
                 }
             } else {
                 return res.status(400).json({
                     success: false,
-                    message: 'Mô tả của món ăn không hợp lệ'
+                    message: 'Tên món ăn không hợp lệ'
                 });
             }
-        } else {
-            return res.status(400).json({
-                success: false,
-                message: 'Tên món ăn không hợp lệ'
-            });
-        }
 
     } else
         return res.status(400).json({
@@ -198,7 +205,13 @@ router.put('/food-edit', verifyAdmin, async (req, res) => {
 router.post('/food-add', verifyAdmin, async (req, res) => {
     const { foodname, foodtype, foodpriceration, fooddescription, foodimagedescription } = req.body
     let foodid
-    if (checkText(foodname)) {
+    if (!checkFoodImage(req.files.foodimage)) {
+        return res.status(400).json({
+            success: false,
+            message: 'Ảnh món ăn không phù hợp, chỉ hỗ trợ định dạng png, jpeg, jpg'
+        });
+    }
+    else if (checkText(foodname)) {
         if (checkText(fooddescription)) {
             try {
                 await new Food().addFood(foodtype.split('LMA')[1], foodname, fooddescription)
